refactor: moved job and network to experimental dir
