+++ src/experimental.ts
@@ -1,3 +1,3 @@
-export * from "./job";
-export * from "./golem_network";
+export * from "./experimental/job";
+export * from "./experimental/golem_network";
 export * from "./experimental/reputation";
