+++ src/index.ts
@@ -8,8 +8,6 @@ export * from "./network";
 export * from "./events";
 export * from "./utils";
 export * from "./utils/yagna/yagna";
-export * from "./golem_network";
-export * from "./job";
 export * from "./work";
 export * from "./error/golem-error";
 export { StatsService } from "./stats/service";
