+++ tests/docker/data-node/ya-provider/globals.json
@@ -1,0 +1,5 @@
+{
+  "node_name": "node-name",
+  "subnet": "public",
+  "account": "0x797cE3Aa8dc255D13E48F31A6B23fe18b5924940"
+}
